:dash: data moved to particular section
