+++ .ipynb_checkpoints/CONTRIBUTING-checkpoint.md
@@ -1,0 +1,35 @@
+| Name              | Github                                                        |
+| ----------------- | ------------------------------------------------------------- |
+| Keshav Singh      | [@keshavsingh4522](https://github.com/keshavsingh4522/)       |
+| Justinus Hermawan | [@justinushermawan](https://github.com/justinushermawan/)     |
+| Shruti Jha        | [@ShrutiJha31](https://github.com/ShrutiJha31/)               |
+| Richard Timothy   | [@RichardTimothy1307](https://github.com/RichardTimothy1307/) |
+| Sumit Soni        | [@isumit19](https://github.com/isumit19/)                     |
+| Siddarth Awasthi  | [@siddarth308](https://github.com/siddarth308/)               |
+| Hero Tiwari       | [@Skywalkers-24](https://github.com/Skywalkers-24/)           |
+| Sarthak Shastri   | [@sarthak815](https://github.com/sarthak815)                  |
+| Aman Gupta        | [@find-aman](https://github.com/find-aman)                    |
+| Vatsal Narwal     | [@VatsalNarwal](https://github.com/VatsalNarwal)              |
+| Pratik Shirodkar  | [@Pratik-Shirodkar](https://github.com/Pratik-Shirodkar)      |
+| Kanishk J         | [@Kanishk7559](https://github.com/kanishk7559)                |
+| Misa Dravid       | [@Misa-Dravid](https://github.com/misa-bot)                   |
+| Pranjal Soni      | [@sonipranjal](https://github.com/sonipranjal)                |
+| Rithvik Ayithapu  | [@rithvik2607](https://github.com/rithvik2607)                |
+| Divy              | [@dr-one-punch](https://github.com/dr-one-punch)              |
+| Juli Cheng        | [@julicheng](https://github.com/julicheng)                    |
+| Dakshay Ahuja     | [@dakshayahuja](https://github.com/dakshayahuja)              |
+| Aman Jain         | [@amanjain](https://github.io/amnjain)                        |
+| Satyam Gadekar    | [@07Satyam](https://github.com/07Satyam)                      |
+| Musadiq Peerzada  | [@muxa11](https://github.com/muxa11)                          |
+| Mayank Sharma     | [@MayankGeek](https://github.com/MayankGeek)                  |
+| Ananya Sajwan     | [@ananya2407](https://github.com/ananya2407)                  |
+| Phagun Jain       | [@PHAGUN-JAIN](https://github.com/PHAGUN-JAIN)                |
+| Deepakshi Sood    | [@deepakshisud](https://github.com/deepakshisud)              |
+| Swayamsidh Nayak  | [@swayam50](https://github.com/swayam50)                      |
+| Arigato haxroot   | [@Arigatohaxroot](https://github.com/Arigatohaxroot)          |
+| Rupali            | [@Rupali409](https://github.com/Rupali409)                    |
+| Muhammad Bilal    | [@Muhammad-Bilal-MB](https://github.com/Muhammad-Bilal-MB)    |
+| Sandi Aris        | [@Sandi-aris](https://github.com/sandi-aris)                  |
+| Misbahud Din      | [@Misbahud-Din](https://github.com/Misbahud-Din)              |
+| Ganzabahl         | [@Ganzabahl](https://github.com/Ganzabahl)                    |
+| Ashish            | [@ashish-garg18](https://github.com/ashish-garg18)            |
