+++ .ipynb_checkpoints/k-checkpoint.cpp
@@ -1,0 +1,7 @@
+#include<iostream>
+using namespace std;
+int main()
+{
+
+return 0;
+}
