+++ .ipynb_checkpoints/p-checkpoint.json
@@ -1,0 +1,167 @@
+{
+    "@keshavsingh4522": {
+        "name": "Keshav Singh",
+        "link": "https://github.com/keshavsingh4522/",
+        "user-pic": "https://avatars.githubusercontent.com/keshavsingh4522"
+    },
+    "@justinushermawan": {
+        "name": "Justinus Hermawan",
+        "link": "https://github.com/justinushermawan/",
+        "user-pic": "https://avatars.githubusercontent.com/justinushermawan"
+    },
+    "@ShrutiJha31": {
+        "name": "Shruti Jha",
+        "link": "https://github.com/ShrutiJha31/",
+        "user-pic": "https://avatars.githubusercontent.com/ShrutiJha31"
+    },
+    "@RichardTimothy1307": {
+        "name": "Richard Timothy",
+        "link": "https://github.com/RichardTimothy1307/",
+        "user-pic": "https://avatars.githubusercontent.com/RichardTimothy1307"
+    },
+    "@isumit19": {
+        "name": "Sumit Soni",
+        "link": "https://github.com/isumit19/",
+        "user-pic": "https://avatars.githubusercontent.com/isumit19"
+    },
+    "@siddarth308": {
+        "name": "Siddarth Awasthi",
+        "link": "https://github.com/siddarth308/",
+        "user-pic": "https://avatars.githubusercontent.com/siddarth308"
+    },
+    "@Skywalkers-24": {
+        "name": "Hero Tiwari",
+        "link": "https://github.com/Skywalkers-24/",
+        "user-pic": "https://avatars.githubusercontent.com/Skywalkers-24"
+    },
+    "@sarthak815": {
+        "name": "Sarthak Shastri",
+        "link": "https://github.com/sarthak815",
+        "user-pic": "https://avatars.githubusercontent.com/sarthak815"
+    },
+    "@find-aman": {
+        "name": "Aman Gupta",
+        "link": "https://github.com/find-aman",
+        "user-pic": "https://avatars.githubusercontent.com/find-aman"
+    },
+    "@VatsalNarwal": {
+        "name": "Vatsal Narwal",
+        "link": "https://github.com/VatsalNarwal",
+        "user-pic": "https://avatars.githubusercontent.com/VatsalNarwal"
+    },
+    "@Pratik-Shirodkar": {
+        "name": "Pratik Shirodkar",
+        "link": "https://github.com/Pratik-Shirodkar",
+        "user-pic": "https://avatars.githubusercontent.com/Pratik-Shirodkar"
+    },
+    "@Kanishk7559": {
+        "name": "Kanishk J",
+        "link": "https://github.com/kanishk7559",
+        "user-pic": "https://avatars.githubusercontent.com/Kanishk7559"
+    },
+    "@Misa-Dravid": {
+        "name": "Misa Dravid",
+        "link": "https://github.com/misa-bot",
+        "user-pic": "https://avatars.githubusercontent.com/Misa-Dravid"
+    },
+    "@sonipranjal": {
+        "name": "Pranjal Soni",
+        "link": "https://github.com/sonipranjal",
+        "user-pic": "https://avatars.githubusercontent.com/sonipranjal"
+    },
+    "@rithvik2607": {
+        "name": "Rithvik Ayithapu",
+        "link": "https://github.com/rithvik2607",
+        "user-pic": "https://avatars.githubusercontent.com/rithvik2607"
+    },
+    "@dr-one-punch": {
+        "name": "Divy",
+        "link": "https://github.com/dr-one-punch",
+        "user-pic": "https://avatars.githubusercontent.com/dr-one-punch"
+    },
+    "@julicheng": {
+        "name": "Juli Cheng",
+        "link": "https://github.com/julicheng",
+        "user-pic": "https://avatars.githubusercontent.com/julicheng"
+    },
+    "@dakshayahuja": {
+        "name": "Dakshay Ahuja",
+        "link": "https://github.com/dakshayahuja",
+        "user-pic": "https://avatars.githubusercontent.com/dakshayahuja"
+    },
+    "@amanjain": {
+        "name": "Aman Jain",
+        "link": "https://github.io/amnjain",
+        "user-pic": "https://avatars.githubusercontent.com/amanjain"
+    },
+    "@07Satyam": {
+        "name": "Satyam Gadekar",
+        "link": "https://github.com/07Satyam",
+        "user-pic": "https://avatars.githubusercontent.com/07Satyam"
+    },
+    "@muxa11": {
+        "name": "Musadiq Peerzada",
+        "link": "https://github.com/muxa11",
+        "user-pic": "https://avatars.githubusercontent.com/muxa11"
+    },
+    "@MayankGeek": {
+        "name": "Mayank Sharma",
+        "link": "https://github.com/MayankGeek",
+        "user-pic": "https://avatars.githubusercontent.com/MayankGeek"
+    },
+    "@ananya2407": {
+        "name": "Ananya Sajwan",
+        "link": "https://github.com/ananya2407",
+        "user-pic": "https://avatars.githubusercontent.com/ananya2407"
+    },
+    "@PHAGUN-JAIN": {
+        "name": "Phagun Jain",
+        "link": "https://github.com/PHAGUN-JAIN",
+        "user-pic": "https://avatars.githubusercontent.com/PHAGUN-JAIN"
+    },
+    "@deepakshisud": {
+        "name": "Deepakshi Sood",
+        "link": "https://github.com/deepakshisud",
+        "user-pic": "https://avatars.githubusercontent.com/deepakshisud"
+    },
+    "@swayam50": {
+        "name": "Swayamsidh Nayak",
+        "link": "https://github.com/swayam50",
+        "user-pic": "https://avatars.githubusercontent.com/swayam50"
+    },
+    "@Arigatohaxroot": {
+        "name": "Arigato haxroot",
+        "link": "https://github.com/Arigatohaxroot",
+        "user-pic": "https://avatars.githubusercontent.com/Arigatohaxroot"
+    },
+    "@Rupali409": {
+        "name": "Rupali",
+        "link": "https://github.com/Rupali409",
+        "user-pic": "https://avatars.githubusercontent.com/Rupali409"
+    },
+    "@Muhammad-Bilal-MB": {
+        "name": "Muhammad Bilal",
+        "link": "https://github.com/Muhammad-Bilal-MB",
+        "user-pic": "https://avatars.githubusercontent.com/Muhammad-Bilal-MB"
+    },
+    "@Sandi-aris": {
+        "name": "Sandi Aris",
+        "link": "https://github.com/sandi-aris",
+        "user-pic": "https://avatars.githubusercontent.com/Sandi-aris"
+    },
+    "@Misbahud-Din": {
+        "name": "Misbahud Din",
+        "link": "https://github.com/Misbahud-Din",
+        "user-pic": "https://avatars.githubusercontent.com/Misbahud-Din"
+    },
+    "@Ganzabahl": {
+        "name": "Ganzabahl",
+        "link": "https://github.com/Ganzabahl",
+        "user-pic": "https://avatars.githubusercontent.com/Ganzabahl"
+    },
+    "@ashish-garg18": {
+        "name": "Ashish",
+        "link": "https://github.com/ashish-garg18",
+        "user-pic": "https://avatars.githubusercontent.com/ashish-garg18"
+    }
+}
