+++ CPP/sum_and_arithmetic_average_of_20_numbers.cpp
