+++ Python/camera.py
